third_party/WebKit/LayoutTests/http/tests/devtools/network/network-domain-filter.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/network-test.js"></script>
<script>
function test() {
  function checkSubdomains(domain) {
    TestRunner.addResult('');
    TestRunner.addResult('Domain: ' + domain);
    TestRunner.addResult('Subdomains: ' + JSON.stringify(Network.NetworkLogView._subdomains(domain)));
  }

  function checkFilter(value, domains) {
    var filter = Network.NetworkLogView._createRequestDomainFilter(value);
    TestRunner.addResult('');
    TestRunner.addResult('Filter: ' + value);
    for (var i = 0; i < domains.length; ++i)
      TestRunner.addResult('Domain \'' + domains[i] + '\' matches: ' + filter({domain: domains[i]}));
  }

  checkSubdomains('foo.bar.com');
  checkSubdomains('thumbnails');

  checkFilter('bar.com', ['foo.bar.com', 'bar.com', 'com']);
  checkFilter('*.bar.com', ['foo.bar.com', 'bar.com']);
  checkFilter('*.bar.*', ['foo.bar.com', 'baz.bar.org', 'bar.foo.net']);

  TestRunner.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>Tests doamin filter.</p>
</body>
</html>
